requirements.txt

# Base
pytz==2019.3
python-slugify==4.0.0
Pillow==7.0.0
django-cleanup==4.0.0
psycopg2-binary==2.8.5

# Django
django==3.0.3

# Environment
django-environ==0.4.5

# Django REST Framework
djangorestframework==3.11.1
django-filter==2.2.0

# JWT
pyjwt>=1.7.1

# Passwords security
argon2-cffi==19.2.0

# Static files
whitenoise==5.0.1

# Celery
redis>=3.2.0
django-redis==4.10.0
celery==4.2.1


# Tools
django-extensions==2.2.9
django-utils-six==2.0
launchpadlib==1.10.13

# Debugging
ipdb==0.11


# Testing
# mypy==0.650
pytest==4.0.2
pytest-sugar==0.9.2
pytest-django==3.4.4
factory-boy==2.11.1
mypy==0.782
# Code quality
flake8==3.6.0

#Stripe
stripe

Faker==4.1.1
django-webpack-loader==0.6.0

gunicorn==20.0.4

# Static files
django-storages[boto3]==1.9.1

# Email
django-anymail[sendgrid]==7.2.1

# Video
moviepy==1.0.3

# Guicorn
django-heroku==0.3.1
gunicorn==20.0.4
